Broader template-format support
You can change the template-renderer by setting auth_template_renderer
to the desired format. It currently defaults to haml.
The option makes some assumptions about how the templates are rendered inside
your app. It is assuming that a method with the name of  the value of
auth_template_renderer is called.